Renamed Modal.svelte > PaymentSection.svelte
